Some minor fix
