+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/UnaryClientStream.java
@@ -69,7 +69,7 @@ public class UnaryClientStream extends AbstractClientStream implements Stream {
         return map;
     }
 
-    private class ClientUnaryInboundTransportObserver extends UnaryInboundTransportObserver implements TransportObserver {
+    private class ClientUnaryInboundTransportObserver extends ServerUnaryInboundTransportObserver implements TransportObserver {
 
         @Override
         public void onComplete() {
